+++ src/main/java/de/christophlorenz/jweatherflux/forwarder/IgnoredConnectionException.java
@@ -1,0 +1,9 @@
+package de.christophlorenz.jweatherflux.forwarder;
+
+/**
+ * Indicates, that we have a connection problem, but want to ignore it
+ */
+public class IgnoredConnectionException extends
+    Throwable {
+
+}